requirements.txt
discord.py==2.3.1
python-dotenv==1.0.0
SQLAlchemy==1.4.39